fetchHandler combined with the FetchMock file
