+++ packages/core/src/Router.js
@@ -1,8 +1,10 @@
 //@type-check
 import Route from './Route.js';
 import { isUrlMatcher, isFunctionMatcher } from './Matchers.js';
+import { buildResponse } from './ResponseBuilder.js';
 /** @typedef {import('./Route').RouteConfig} RouteConfig */
 /** @typedef {import('./Route').RouteResponse} RouteResponse */
+/** @typedef {import('./Route').RouteResponseFunction} RouteResponseFunction */
 /** @typedef {import('./Matchers').RouteMatcher} RouteMatcher */
 /** @typedef {import('./FetchMock').FetchMockConfig} FetchMockConfig */
 /** @typedef {import('./FetchMock')} FetchMock */
@@ -18,6 +20,79 @@ import { isUrlMatcher, isFunctionMatcher } from './Matchers.js';
 const nameToOptions = (options) =>
 	typeof options === 'string' ? { name: options } : options;
 
+/**
+ * 
+ * @param {RouteResponse} response 
+ * @returns {RouteResponse is RouteResponseFunction}
+ */
+const isPromise = response => typeof /** @type {Promise<any>} */(response).then === 'function'
+
+/**
+ * @param {RouteResponse} response
+ * @param {string} url
+ * @param {RequestInit} options
+ * @param {Request} request
+ * @returns
+ */
+const resolveUntilResponseConfig = async (
+	response,
+	url,
+	options,
+	request,
+) => {
+	// We want to allow things like
+	// - function returning a Promise for a response
+	// - delaying (using a timeout Promise) a function's execution to generate
+	//   a response
+	// Because of this we can't safely check for function before Promisey-ness,
+	// or vice versa. So to keep it DRY, and flexible, we keep trying until we
+	// have something that looks like neither Promise nor function
+	//eslint-disable-next-line no-constant-condition
+	while (true) {
+		if (typeof response === 'function') {
+			response = response(url, options, request);
+		} else if (isPromise(response)) {
+			response = await response; // eslint-disable-line  no-await-in-loop
+		} else {
+			return response;
+		}
+	}
+};
+
+/**
+ *
+ * @param {Object} input
+ * @param {Route} input.route
+ * @param {string} input.url
+ * @param {RequestInit} input.options
+ * @param {Request} [input.request]
+ * @param {CallLog} input.callLog
+ * @returns {Promise<Response>}
+ */
+const generateResponse = async ({
+	route,
+	url,
+	options,
+	request,
+	callLog,
+}) => {
+	const response = await resolveUntilResponseConfig(
+		route.config.response,
+		url,
+		options,
+		request,
+	);
+
+	const [realResponse, finalResponse] = buildResponse({
+		url,
+		responseConfig: response,
+		route,
+	});
+
+	callLog.response = realResponse;
+
+	return finalResponse;
+};
 
 export default class Router {
 	/**
@@ -39,7 +114,7 @@ export default class Router {
 
 	/**
 	 * @param {NormalizedRequest} normalizedRequest
-	 * @returns {{route: Route , callLog: CallLog}}
+	 * @returns {{route: Route , callLog: CallLog, response: Promise<Response>}}
 	 */
 	execute({ url, options, request }) {
 		const routesToTry = this.fallbackRoute ? [...this.routes, this.fallbackRoute]: this.routes 
@@ -48,15 +123,26 @@ export default class Router {
 		);
 
 		if (route) {
-			return {
+			const callLog = {
+				url,
+				options,
+				request,
 				route,
-				callLog: {
-					url,
-					options,
-					request,
-					route,
-				},
+			}
+			const response = generateResponse({
+				route,
+				url,
+				options,
+				request,
+				callLog,
+			});
+			return {
+				response,
+				route,
+				callLog,
 			};
+
+			
 		}
 
 		throw new Error(
